[Refactoring] Tab bar and give user manager as an env object
