+++ Naenio/Sources/Home/HomeView.swift
@@ -10,6 +10,7 @@ import Combine
 import Introspect
 
 struct HomeView: View {
+    @EnvironmentObject var userManager: UserManager
     @StateObject var viewModel = HomeViewModel()
     @ObservedObject var scrollViewHelper = ScrollViewHelper()
 
@@ -132,6 +133,7 @@ struct HomeView: View {
             }
             .sheet(isPresented: $showComments) {
                 CommentView(isPresented: $showComments, parentId: $selectedPostId)
+                    .environmentObject(userManager)
             }
         }
