+++ Naenio/Sources/Main/MainView.swift
@@ -6,24 +6,68 @@
 //
 
 import SwiftUI
+import Combine
 
 struct MainView: View {
     @EnvironmentObject var userManager: UserManager
     
-    @State var tabBarPages: [TabBarPage] = [
+    @State var selectedTab = 1
+    @State fileprivate var tabBarLowSheetInfo = TabBarLowSheetInfo(isPresented: false, postId: 0)
+    
+    @State var pages: [TabBarPage] = [
         TabBarPage(pageName: .curation, selectedIcon: "tab_curation_selected", deselectedIcon: "tab_curation_deselected", tag: 0),
         TabBarPage(pageName: .home, selectedIcon: "tab_home_selected", deselectedIcon: "tab_home_deselected", tag: 1),
         TabBarPage(pageName: .myPage, selectedIcon: "tab_myPage_selected", deselectedIcon: "tab_myPage_deselected", tag: 2)
     ]
     
     var body: some View {
-        TabBarView(pages: $tabBarPages)
-            .background(Color.background.ignoresSafeArea())
+        TabView(selection: $selectedTab) {
+            ForEach(pages) { item in
+                TabContentView(pageName: item.pageName)
+                    .environmentObject(userManager)
+                    .tabItem {
+                        Image(self.selectedTab == item.tag ? item.selectedIcon : item.deselectedIcon)
+                        Text("") // For padding
+                    }
+                    .tag(item.tag)
+            }
+        }
+        .introspectTabBarController { controller in
+            controller.tabBar.backgroundColor = UIColor(Color.tabBarBackground)
+            controller.tabBar.shadowImage = UIImage()
+            controller.tabBar.backgroundImage = UIImage()
+            
+            controller.tabBar.layer.masksToBounds = true
+            controller.tabBar.layer.cornerRadius = 14
+            controller.tabBar.layer.maskedCorners = [.layerMinXMinYCorner, .layerMaxXMinYCorner]
+        }
+        .onReceive(Publishers.lowSheetNotificationPublisher) { (value: LowSheetNotification) in
+            tabBarLowSheetInfo.isPresented = true
+            tabBarLowSheetInfo.postId = value.postId
+        }
+        .lowSheet(isPresented: $tabBarLowSheetInfo.isPresented) {
+            ReportAndShareSheetView(
+                isPresented: $tabBarLowSheetInfo.isPresented,
+                postID: tabBarLowSheetInfo.postId
+            )
+                .padding(.horizontal, 27)
+        }
     }
 }
 
-struct MainView_Previews: PreviewProvider {
-    static var previews: some View {
-        MainView()
+fileprivate struct TabBarLowSheetInfo {
+    var isPresented: Bool
+    var postId: Int
+}
+
+struct TabBarPage: Identifiable {
+    var id = UUID()
+    var pageName: PageType
+    var selectedIcon: String
+    var deselectedIcon: String
+    var tag: Int
+    
+    enum PageType {
+        case curation, home, myPage
     }
 }
